Revert "Home-screen code clean"
