+++ css/main.css
@@ -26,7 +26,6 @@ html {
   font-family: var(--font-primary), var(--font-fallback) !important;
   line-height: 1.5;
   background-color: var(--white);
-  scroll-behavior: smooth;
 }
 body,
 h1,
@@ -148,7 +147,6 @@ main {
   text-decoration: none;
   pointer-events: none;
 }
-
 .menu-icon {
   --color: var(--black);
   width: 1.75rem;
